IMRT-414: separate sync job config into its own file
