+++ src/main/java/org/opentestsystem/ap/imrt/iis/config/ItemSynchronizationBatchJobConfiguration.java
@@ -1,0 +1,97 @@
+package org.opentestsystem.ap.imrt.iis.config;
+
+import org.opentestsystem.ap.imrt.iis.batch.ItemSynchronizationJobExecutionListener;
+import org.opentestsystem.ap.imrt.iis.batch.ItemSynchronizationTasklet;
+import org.opentestsystem.ap.imrt.iis.service.ItemSynchronizationService;
+import org.springframework.batch.core.Job;
+import org.springframework.batch.core.JobExecutionListener;
+import org.springframework.batch.core.Step;
+import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
+import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
+import org.springframework.batch.core.explore.JobExplorer;
+import org.springframework.batch.core.launch.support.RunIdIncrementer;
+import org.springframework.batch.core.step.tasklet.Tasklet;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Configure Spring Batch to execute the item synchronization process as a {@link org.springframework.batch.core.Job}.
+ * <p>
+ * Item Synchronization Job rules:
+ * </p>
+ * <p>
+ * 1. Only one instance of the job should run at a time. If the job is currently running, other requests to start the
+ * item synchronization job should be rejected
+ * 2. The job should not run when Spring Boot starts up. Instead, the job should be started by a call to a web endpoint
+ * (e.g. a RESTful call)
+ * 3. The same job should be able to run multiple times (i.e. it should be repeatable).  That is, the Item
+ * Synchronization Process should be able to run once a day (or at some other regular interval)
+ * </p>
+ */
+@Configuration
+public class ItemSynchronizationBatchJobConfiguration {
+    /**
+     * Create a {@link org.springframework.batch.core.Job} that complies with the rules cited above.
+     *
+     * @param itemSynchronizationStep A {@link org.springframework.batch.core.Step} that executes the item
+     *                                synchronization process
+     * @return A {@link org.springframework.batch.core.Job} with the step(s) necessary to execute the item
+     * synchronization process.
+     */
+    @Bean
+    public Job itemSynchronizationJob(final JobBuilderFactory jobBuilderFactory,
+                                      final Step itemSynchronizationStep,
+                                      final JobExplorer jobExplorer) {
+        return jobBuilderFactory.get("itemSynchronizationJob")
+                .incrementer(new RunIdIncrementer())
+                .listener(itemSynchronizationJobExecutionListener(jobExplorer))
+                .start(itemSynchronizationStep)
+                .build();
+    }
+
+    /**
+     * A {@link org.springframework.batch.core.Step} to execute the item synchronization process within a
+     * {@link org.springframework.batch.core.Job}.
+     *
+     * @param itemSynchronizationTasklet The {@link org.springframework.batch.core.step.tasklet.Tasklet} that rune the
+     *                                   item synchronization process.
+     * @return A {@link org.springframework.batch.core.Step} for the item synchronization process
+     * {@link org.springframework.batch.core.Job}
+     */
+    @Bean
+    public Step itemSynchronizationStep(final StepBuilderFactory stepBuilderFactory,
+                                        final Tasklet itemSynchronizationTasklet) {
+        return stepBuilderFactory.get("itemSynchronizationStep")
+                .tasklet(itemSynchronizationTasklet)
+                .allowStartIfComplete(true)
+                .build();
+    }
+
+    /**
+     * A {@link org.springframework.batch.core.step.tasklet.Tasklet} representing the item synchronization process.
+     * This is what the {@link org.springframework.batch.core.Step} will call to do the actual item synchronization
+     * work.
+     *
+     * @param itemSynchronizationService An implementation of the
+     *                                   {@link org.opentestsystem.ap.imrt.iis.service.ItemSynchronizationService} to
+     *                                   perform the synchronization process.
+     * @return A {@link org.springframework.batch.core.step.tasklet.Tasklet} implementation that can execute the item
+     * synchronization process
+     */
+    @Bean
+    public ItemSynchronizationTasklet itemSynchronizationTasklet(final ItemSynchronizationService itemSynchronizationService) {
+        return new ItemSynchronizationTasklet(itemSynchronizationService);
+    }
+
+    /**
+     * A {@link org.springframework.batch.core.JobExecutionListener} implementation that will ensure only one instance
+     * of the item synchronization process {@link org.springframework.batch.core.Job} is running.
+     *
+     * @return A {@link org.springframework.batch.core.JobExecutionListener} implementation that prevents running
+     * multiple instances of the item synchronization process {@link org.springframework.batch.core.Job}s.
+     */
+    @Bean
+    public JobExecutionListener itemSynchronizationJobExecutionListener(final JobExplorer jobExplorer) {
+        return new ItemSynchronizationJobExecutionListener(jobExplorer);
+    }
+}
